+++ launchers/default.sh
@@ -13,7 +13,7 @@ dt-launchfile-init
 # NOTE: Use `dt-exec COMMAND` to run the main process (blocking process)
 
 # launching app
-roslaunch perception perception.launch
+roslaunch perception perception.launch veh:=$VEHICLE_NAME
 # rosrun controls controls.launch
