+++ packages/as_msgs/msg/WheelOdometry.msg
@@ -1,6 +1,3 @@
 Header header
-float32 x
-float32 y
-float32 heading
 float32 left_wheel_velocity
 float32 right_wheel_velocity
