+++ packages/perception/CMakeLists.txt
@@ -3,6 +3,7 @@ project(perception)
 
 find_package(catkin REQUIRED COMPONENTS
   rospy
+  std_msgs
 )
 
 catkin_package()
