+++ packages/perception/package.xml
@@ -8,4 +8,6 @@
   <license>None</license>
 
   <buildtool_depend>catkin</buildtool_depend>
+
+  <run_depend>as_msgs</run_depend>
 </package>
